FIX create pedido and create default registers in main comments

+++ Mozo/src/main/java/com/incade/poo/mozo/dto/PedidoDto.java
@@ -2,6 +2,6 @@ package com.incade.poo.mozo.dto;
 
 import java.util.List;
 
-public record PedidoDto(Long id, Long mesa, MozoDto mozo, String estado, Double total, List<ItemDto> items) {
+public record PedidoDto(Long id, Integer mesa, MozoDto mozo, String estado, Double total, List<ItemDto> items) {
 
 }
